requirements.txt
flask
tensorflow-cpu
opencv-python
numpy
pytesseract
keras-ocr
gunicorn
